Move engineDoFinal to subclasses.
